refactor and remove deadcode
