+++ crates/burn-jit/src/kernel/reduce/mod.rs
@@ -1,12 +1,5 @@
 mod base;
-// mod naive;
-// mod prod;
-// mod shared;
-// mod subcube;
-// mod sum;
-// mod tune;
+mod tune;
 
 pub use base::*;
-// pub use prod::*;
-// pub use sum::*;
-// pub use tune::*;
+pub use tune::*;
