--- crates/burn-jit/src/kernel/reduce/naive/base.rs
@@ -1,25 +1,0 @@
-use cubecl::prelude::*;
-
-pub trait ReduceDimNaiveFamily: Send + Sync + 'static {
-    type Reduce<E: Numeric>: ReduceDimNaive<E>;
-}
-
-/// Specifies the reduce dim algorithm in use
-#[cube]
-pub trait ReduceDimNaive<EI: Numeric>: Send + Sync + 'static {
-    /// The reduction accumulator
-    type Accumulator: CubeType;
-
-    /// Initialization for naive algorithm
-    fn initialize_naive() -> Self::Accumulator;
-
-    /// Inner loop for naive algorithm
-    fn inner_loop_naive(accumulator: &mut Self::Accumulator, current_value: EI, i: u32);
-
-    /// Assignation for naive algorithm
-    fn assign_naive<EO: Numeric>(
-        output: &mut Tensor<EO>,
-        accumulator: Self::Accumulator,
-        shape_reduce_dim: u32,
-    );
-}
