--- crates/burn-jit/src/kernel/reduce/subcube/mod.rs
@@ -1,7 +1,0 @@
-pub mod argmax;
-pub mod argmin;
-pub mod base;
-pub mod kernel;
-pub mod mean_dim;
-pub mod prod_dim;
-pub mod sum_dim;
