--- crates/burn-jit/src/kernel/reduce/tune/mod.rs
@@ -1,7 +1,0 @@
-#[cfg(feature = "autotune")]
-mod base;
-mod key;
-
-#[cfg(feature = "autotune")]
-pub(crate) use base::*;
-pub use key::*;
